fix delete

+++ modules/07 NodeJS/Assignments/05/chat/src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component, ViewChild, ElementRef } from '@angular/core';
-import { FormBuilder } from '@angular/forms';
+import { FormBuilder, Validators } from '@angular/forms';
 import { ChatService } from './chat.service';
 import { chat } from './chat.model';
 
@@ -12,10 +12,10 @@ export class AppComponent {
   chat: chat[];
   @ViewChild('chatName') chatName: ElementRef;
   @ViewChild('chatMessage') chatMessage: ElementRef;
-  @ViewChild('MessageId') MessageId : ElementRef;
+  @ViewChild('MessageId') MessageId: ElementRef;
   chatForm = this.fb.group({
-    chatMessage: [''],
-    chatName: ['']
+    chatMessage: ['', Validators.required],
+    chatName: ['', Validators.required]
   });
 
   constructor(private fb: FormBuilder, private chatService: ChatService) {
@@ -26,13 +26,18 @@ export class AppComponent {
     this.chatService.getMessages().then(data => this.chat = data.result);
   }
   postMessages() {
-    this.chatService.postMessage(this.chatName.nativeElement.value, this.chatMessage.nativeElement.value);
+    if (this.chatForm.valid) {
+      this.chatService.postMessage(this.chatName.nativeElement.value, this.chatMessage.nativeElement.value);
+    }
   }
   deleteMessage() {
-    this.chatService.deleteMessage(this.MessageId.nativeElement.value);
-
+    if (this.chatForm.valid) {
+      this.chatService.deleteMessage(this.MessageId.nativeElement.value);
+    }
   }
   updateMessage() {
-    this.chatService.updateMessage(this.MessageId.nativeElement.value , this.chatMessage.nativeElement.value);
+    if (this.chatForm.valid) {
+      this.chatService.updateMessage(this.MessageId.nativeElement.value, this.chatMessage.nativeElement.value);
+    }
   }
 }
